jose.messa/ADDED Register user from login page
