+++ src/css/style.css
@@ -11,6 +11,10 @@ html {
 a {
   color: inherit;
   text-decoration: none;
+  cursor: pointer;
+}
+button{
+  cursor: pointer;
 }
 body{
   position: relative;
