+++ src/css/users-list.css
@@ -8,6 +8,7 @@
 
   overflow: hidden;
   background-color: rgb(255, 255, 255);
+  cursor: pointer;
   
 }
